EXPERIMENTAL: master-only jenkins deployments
This experimental commit allows for master-only Jenkins deployments.
Instead of having Jenkins masters (running on Mesos) provision slaves
on-demand (also running on Mesos), it shifts the ownership of the
executors to the masters. This allows for easier global queueing (based
on threads instead of resources) and only requires scaling the masters.

It's not yet known what the implications of this would be, but it might
be worth exploring. Or this branch might be a waste of time. We'll find
out.